Projeto Your New Car Concluído

+++ yournewcar/src/main/java/br/com/andersonchoren/yournewcar/service/CarService.java
@@ -1,13 +1,11 @@
 package br.com.andersonchoren.yournewcar.service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import br.com.andersonchoren.yournewcar.model.Car;
 import br.com.andersonchoren.yournewcar.model.dto.CarDTO;
 import br.com.andersonchoren.yournewcar.repository.CarRepository;
 import br.com.andersonchoren.yournewcar.util.CarMapping;
@@ -21,10 +19,6 @@ public class CarService {
         var result = carRepository.save(CarMapping.toCar(car));
         return CarMapping.fromCar(result);
     }
-
-    // public Car insertOrUpdate(Car car) {
-    // return carRepository.save(car);
-    // }
 
     public List<CarDTO> findAll() {
         return carRepository.findAll().stream().map(car -> CarMapping.fromCar(car)).collect(Collectors.toList());
@@ -40,10 +34,14 @@ public class CarService {
     }
 
     public boolean delete(int id) {
-        carRepository.deleteById(id);
-
         var result = carRepository.findById(id);
-        return result.isEmpty();
+        if(result.isPresent()){
+            carRepository.deleteById(id);
+            result = carRepository.findById(id);
+            return result.isEmpty();
+        }
+        else
+            return false;
     }
 
 }
